refactor: package 정리 및 깨지는 테스트 수정
