--- src/main/java/http/exception/HttpHeaderRegistrationException.java
@@ -1,11 +1,0 @@
-package http.exception;
-
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-public class HttpHeaderRegistrationException extends IllegalArgumentException {
-
-    public HttpHeaderRegistrationException(String s) {
-        super(s);
-    }
-}
