--- src/main/java/http/method/HttpMethod.java
@@ -1,9 +1,0 @@
-package http.method;
-
-public enum HttpMethod {
-
-    GET,
-    POST,
-    PUT,
-    DELETE
-}
