--- src/main/java/http/querystring/exception/QueryStringParsingException.java
@@ -1,11 +1,0 @@
-package http.querystring.exception;
-
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-public class QueryStringParsingException extends IllegalArgumentException {
-
-    public QueryStringParsingException(String s) {
-        super(s);
-    }
-}
