+++ src/main/java/http/request/HttpHeaders.java
@@ -1,0 +1,17 @@
+package http.request;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class HttpHeaders {
+
+    private final Map<String, String> headers = new HashMap<>();
+
+    public String get(String headerKey) {
+        return headers.get(headerKey);
+    }
+
+    public void put(String key, String value) {
+        headers.put(key, value);
+    }
+}
