+++ src/main/java/http/request/exception/HttpHeaderRegistrationException.java
@@ -1,0 +1,11 @@
+package http.request.exception;
+
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class HttpHeaderRegistrationException extends IllegalArgumentException {
+
+    public HttpHeaderRegistrationException(String s) {
+        super(s);
+    }
+}
