+++ src/main/java/http/request/method/HttpMethod.java
@@ -1,0 +1,9 @@
+package http.request.method;
+
+public enum HttpMethod {
+
+    GET,
+    POST,
+    PUT,
+    DELETE
+}
