+++ src/main/java/http/request/querystring/exception/QueryStringParsingException.java
@@ -1,0 +1,11 @@
+package http.request.querystring.exception;
+
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class QueryStringParsingException extends IllegalArgumentException {
+
+    public QueryStringParsingException(String s) {
+        super(s);
+    }
+}
