+++ src/main/java/http/request/requestline/exception/RequestLineParsingException.java
@@ -1,0 +1,11 @@
+package http.request.requestline.exception;
+
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class RequestLineParsingException extends IllegalArgumentException {
+
+    public RequestLineParsingException(String s) {
+        super(s);
+    }
+}
