+++ src/main/java/http/request/requestline/path/exception/FileExtensionNotFoundException.java
@@ -1,0 +1,4 @@
+package http.request.requestline.path.exception;
+
+public class FileExtensionNotFoundException extends RuntimeException {
+}
