+++ src/main/java/http/request/requestline/protocol/Protocol.java
@@ -1,0 +1,5 @@
+package http.request.requestline.protocol;
+
+public enum Protocol {
+    HTTP
+}
