--- src/main/java/http/requestline/exception/RequestLineParsingException.java
@@ -1,11 +1,0 @@
-package http.requestline.exception;
-
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-public class RequestLineParsingException extends IllegalArgumentException {
-
-    public RequestLineParsingException(String s) {
-        super(s);
-    }
-}
