--- src/main/java/http/requestline/path/exception/FileExtensionNotFoundException.java
@@ -1,4 +1,0 @@
-package http.requestline.path.exception;
-
-public class FileExtensionNotFoundException extends RuntimeException {
-}
