--- src/main/java/http/requestline/protocol/Protocol.java
@@ -1,5 +1,0 @@
-package http.requestline.protocol;
-
-public enum Protocol {
-    HTTP
-}
